server/requirements.txt
sanic
configparser
sqlalchemy
selenium
requests
pyjwt
psycopg2-binary
undetected_chromedriver
pytest
pytest-asyncio
httpx
sanic-testing